+++ src/squlearn/__init__.py
@@ -1,12 +1,12 @@
 """A library for quantum machine learning following the scikit-learnstandard."""
 from .util import Executor
-from . import observable, feature_map, kernel, optimizers, qnn, util
+from . import observables, feature_map, kernel, optimizers, qnn, util
 
 __version__ = "0.3.0"
 
 __all__ = [
     "Executor",
-    "observable",
+    "observables",
     "feature_map",
     "kernel",
     "optimizers",
