+++ src/squlearn/observables/observable_base.py
