+++ src/squlearn/observables/observable_implemented/custom_observable.py
