+++ src/squlearn/observables/observable_implemented/ising_hamiltonian.py
