+++ src/squlearn/observables/observable_implemented/single_pauli.py
