+++ src/squlearn/observables/observable_implemented/single_probability.py
