+++ src/squlearn/observables/observable_implemented/summed_paulis.py
