+++ src/squlearn/observables/observable_implemented/summed_probabilities.py
