+++ src/squlearn/qnn/qnn.py
@@ -5,8 +5,8 @@ from qiskit import QuantumCircuit
 from qiskit.circuit import ParameterVector, ParameterExpression
 from qiskit.circuit.parametervector import ParameterVectorElement
 
-from ..observable.observable_base import ObservableBase
-from ..observable.observable_derivatives import (
+from ..observables.observable_base import ObservableBase
+from ..observables.observable_derivatives import (
     ObservableDerivatives,
 )
