+++ tests/qnn/test_qnnc.py
@@ -6,7 +6,7 @@ from sklearn.datasets import make_blobs
 from sklearn.preprocessing import LabelBinarizer, MinMaxScaler
 
 from squlearn import Executor
-from squlearn.observable import SummedPaulis
+from squlearn.observables import SummedPaulis
 from squlearn.feature_map import ChebPQC
 from squlearn.optimizers import SLSQP, Adam
 from squlearn.qnn import QNNClassifier, SquaredLoss
