+++ tests/qnn/test_training.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from squlearn import Executor
-from squlearn.observable import SummedPaulis
+from squlearn.observables import SummedPaulis
 from squlearn.feature_map import ChebPQC
 from squlearn.optimizers import SLSQP
 from squlearn.qnn.loss import SquaredLoss
